HPW-1762 COD Anticipado 2024-12-26 Oscar Rodriguez ADD Se agregaron validaciones para realizar el porcentaje de devolucion decimal

+++ DeliveryBackOffice(DEV)/dbo/Tables/RateHeader.sql
@@ -31,7 +31,7 @@ CREATE TABLE [dbo].[RateHeader] (
     [PackagesRangeId]      INT             NULL,
     [IdCurrency]           INT             NULL,
     [GuideAmountCOD]       DECIMAL(18,2)   NULL,
-    [ReturnPercent]        INT             NULL,
+    [ReturnPercent]        DECIMAL(18,2)   NULL,
     [IsOldest]             INT             NULL,
     [MinGuidesPerMonth]    INT             NULL,
     PRIMARY KEY CLUSTERED ([RheId] ASC),
